Data comunication remote implemented

+++ app/src/main/java/com/juansanz/themovieapp/ui/screens/Navigation.kt
@@ -15,7 +15,7 @@ sealed class Screen(
     object Main : Screen("main")
 
     object Detail : Screen("detail/{${NavArgs.ItemId.key}}") {
-        fun createRoute(id: Int) = "detail/$id"
+        fun createRoute(id: Long) = "detail/$id"
     }
 }
